% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayToGdx.R
\name{arrayToGdx}
\alias{arrayToGdx}
\title{Prepare data frame for GDX}
\usage{
arrayToGdx(a, name)
}
\arguments{
\item{a}{array with factor columns for gams dimensions and one value
column.}

\item{name}{character, object name}
}
\value{
A list with a particular structure required by [gdxrrw::wgdx()].
}
\description{
Create an object that can be handled by [gdxrrw::wgdx()] to write the data
frame as a gams parameter into a gdx file.
}
\details{
This function takes an array and creates a list ready to be written into a
gdx file. It will currently always create an output for a gams parameter.
}
\author{
Robin Hasse
}
