% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStockBal.R
\name{checkStockBal}
\alias{checkStockBal}
\title{Check if stock balance is fulfilled}
\usage{
checkStockBal(path)
}
\arguments{
\item{path}{character, directory of a brick run}
}
\value{
named list with deviations of each balance equation
}
\description{
Only works for sequential renovation right now.
}
\author{
Robin Hasse
}
