% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeShareRen.R
\name{computeShareRen}
\alias{computeShareRen}
\title{Compute the share of heating systems to be renovated}
\usage{
computeShareRen(
  params,
  m,
  cutOffShare = 1,
  standingLifeTime = FALSE,
  timePeriods = NULL,
  vintages = NULL
)
}
\arguments{
\item{params}{data frame with parameters of the Weibull distribution}

\item{m}{Gams container with at least \code{ttotIn}, \code{ttotOut}, \code{tinit} and \code{p_dt}}

\item{cutOffShare}{numeric, renovation share at which we cut off the distribution function by jumping to 1}

\item{standingLifeTime}{logical, whether standing lifetime needs to be considered}

\item{timePeriods}{data frame with time periods for which the calculations should be done}

\item{vintages}{data frame with vintages and their starting and end year}
}
\value{
Data frame of shares to be renovated in each time period
}
\description{
Calculate share of buildings that need to be renovated or demolished between
given time steps assuming a Weibull distribution of the technology life time.
}
\details{
If \code{standingLifeTime = TRUE}, a standing stock is considered.
For most technologies we assume a constant inflow between the start of the
respective vintage cohort and initial time.
For the more recent technology of heat pumps (\code{ehp1}), we assume that
the inflow decreases quadratically with the age of the system going back from
initial time to the start of the vintage cohort, but at most 25 years.
For the rather outdated technologies solids (\code{sobo}) and liquids (\code{libo}) boilers,
we assume that the inlow increases quadratically with the age of the system going back from
initial time to the start of the vintage cohort, but at most 25 (\code{sobo}) or 20 (\code{libo}) years.
}
