% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyGamsFiles.R
\name{copyGamsFiles}
\alias{copyGamsFiles}
\title{Copy gams scripts to output folder}
\usage{
copyGamsFiles(path, overwrite = FALSE)
}
\arguments{
\item{path}{character vector with folders to write input data into}

\item{overwrite}{logical, should existing input.gdx be overwritten?}
}
\description{
Copy gams scripts to output folder
}
\author{
Robin Hasse
}
