% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyGdxFile.R
\name{copyGdxFile}
\alias{copyGdxFile}
\title{Copy GDX file}
\usage{
copyGdxFile(copyFile, pasteDir, pasteFileName, overwrite = FALSE)
}
\arguments{
\item{copyFile}{character, path to origin file}

\item{pasteDir}{character, path to target directory}

\item{pasteFileName}{character, target file name}

\item{overwrite}{logical, should existing input.gdx be overwritten?}
}
\description{
Copy GDX file
}
\author{
Robin Hasse
}
