% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMatchingData.R
\name{createMatchingData}
\alias{createMatchingData}
\title{Create data for reference matching}
\usage{
createMatchingData(path, config, overwrite = FALSE)
}
\arguments{
\item{path}{character vector with folders to run the model in}

\item{config}{run configurations}

\item{overwrite}{logical, should existing data be overwritten?}
}
\description{
This function also performs auto gams code generation to load the reference
sets and parameters
}
\author{
Robin Hasse
}
