% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParameters.R
\name{createParameters}
\alias{createParameters}
\title{Create parameters}
\usage{
createParameters(m, config, inputDir)
}
\arguments{
\item{m}{gams Container, central object to store all data for input.gdx}

\item{config}{named list with run configuration}

\item{inputDir}{directory of input folder}
}
\value{
gams Container with parameters added
}
\description{
Add all parameters to gams container based on config
}
\author{
Robin Hasse
}
