% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRunFolder.R
\name{createRunFolder}
\alias{createRunFolder}
\title{Create new run folder}
\usage{
createRunFolder(
  path,
  config = NULL,
  overwrite = FALSE,
  recursive = FALSE,
  showWarnings = TRUE
)
}
\arguments{
\item{path}{character vector, containing the path(s) to the folder(s) to be created.}

\item{config}{list with run configuration}

\item{overwrite}{logical, should existing folders be overwritten?}

\item{recursive}{logical, should elements of the path other than the last be created?}

\item{showWarnings}{logical, should a warning on not created paths be shown?}
}
\description{
Create a folder for the model to run in and copy required gams files there.
}
\author{
Robin Hasse
}
