% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSets.R
\name{createSets}
\alias{createSets}
\title{Create sets}
\usage{
createSets(m, config)
}
\arguments{
\item{m}{gams Container, central object to store all data for input.gdx}

\item{config}{named list with run configuration}
}
\value{
gams Container with sets added
}
\description{
Add all sets to gams container based on config
}
\author{
Robin Hasse
}
