% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfToGdx.R
\name{dfToGdx}
\alias{dfToGdx}
\title{Prepare data frame for GDX}
\usage{
dfToGdx(
  df,
  name,
  type = "parameter",
  uelCols = NULL,
  valcol = "value",
  field = "l"
)
}
\arguments{
\item{df}{data.frame with factor columns for gams dimensions and one value
column.}

\item{name}{character, object name}

\item{type}{character, gams object type}

\item{uelCols}{character vector, column names of dimension columns, if NULL,
all columns but the value column are considered.}

\item{field}{character, for `valCol = "variable"`, it indicates which field
is given as values.}

\item{valCol}{character, column name of the value column.}
}
\value{
A list with a particular structure required by [gdxrrw::wgdx()].
}
\description{
Create an object that can be handled by [gdxrrw::wgdx()] to write the data
frame as a gams object into a gdx file.
}
\author{
Robin Hasse
}
