% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.addSpecCostToInput}
\alias{.addSpecCostToInput}
\title{Write the specific costs to the input.gdx}
\usage{
.addSpecCostToInput(
  m,
  path,
  optimVar,
  xinit,
  tcalib,
  dims,
  vinExists,
  vinCalib,
  varName = "x",
  shiftIntang = TRUE
)
}
\arguments{
\item{m}{Gams transfer container with previous input data}

\item{path}{character, path to this run}

\item{optimVar}{list of data frames with optimization variables of all flows}

\item{xinit}{list of data frames with initial specific intangible costs of all flows}

\item{tcalib}{numeric, calibration time steps}

\item{dims}{list of characters, dimensions of each flow}

\item{vinExists}{data frame of vintages that exist for each time period}

\item{vinCalib}{data frame with vintages that exist in calibration periods}

\item{varName}{character, optimization variable to calculate specific costs from.
Needs to be a column of optimVarCon and optimVarRen, should be one of 'x', 'xA' or 'xMin'.}

\item{shiftIntang}{logical indicating whether intangible costs should be shifted to positive range}
}
\description{
Write the specific costs to the input.gdx
}
