% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.addTargetsToInput}
\alias{.addTargetsToInput}
\title{Add the calibration targets to the input gdx}
\usage{
.addTargetsToInput(mInput, path, p_calibTarget, dims)
}
\arguments{
\item{mInput}{gamstransfer container of the input gdx}

\item{path}{character, path to output folder of this run}

\item{p_calibTarget}{list of data frames of calibration targets}

\item{dims}{list of characters, dimensions of data}
}
\value{
gamstransfer container with the modified input gdx (invisibly)
}
\description{
Add the calibration targets to the input gdx
}
