% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.addTimeStamp}
\alias{.addTimeStamp}
\title{Add time stamp}
\usage{
.addTimeStamp(x)
}
\arguments{
\item{x}{character}
}
\value{
character with time stamp suffix
}
\description{
Append current date and time
}
