% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.adjustStepSizeAfterArmijo}
\alias{.adjustStepSizeAfterArmijo}
\title{Handle the case that the step size adaptation condition is not satisfied after the predefined number of iterations.}
\usage{
.adjustStepSizeAfterArmijo(armijoStep, stepSizeParams)
}
\arguments{
\item{armijoStep}{data frame with combinations for which the condition is not satisfied}

\item{stepSizeParams}{data frame with the parameters of the step size adaptation procedure}
}
\description{
If for any combination the condition is still not satisfied:
Use the step size with the minimum objective function.
Print a warning if this minimum does not exist set step size to zero.
}
