% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.adjustStepSizeAfterIteration}
\alias{.adjustStepSizeAfterIteration}
\title{Handle the case that the step size adaptation condition is not satisfied after the predefined number of iterations.}
\usage{
.adjustStepSizeAfterIteration(totalStep, stepSizeParams)
}
\arguments{
\item{totalStep}{data frame with combinations for which the condition is not satisfied}

\item{stepSizeParams}{data frame with the parameters of the step size adaptation procedure}
}
\value{
data frame with step size and related paramters, updated if applicable.
}
\description{
If for any combination the condition is still not satisfied:
Set the step size to zero and print a warning if no descent has been detected.
Use the last step size of the iteration otherwise.
}
