% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.aggregateDim}
\alias{.aggregateDim}
\title{Aggregate data across given dimensions by a given function}
\usage{
.aggregateDim(df, agg, func = sum, valueNames = "value")
}
\arguments{
\item{df}{data frame with data to be aggregated}

\item{agg}{character, columns with the dimensions to be aggregated}

\item{func}{function to be used for aggregation}

\item{valueNames}{character, names of the columns containing the values to aggregate}
}
\value{
data frame with aggregated data
}
\description{
Aggregate data across given dimensions by a given function
}
