% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMadratData.R
\name{.argsHash}
\alias{.argsHash}
\title{Get hash for further arguments}
\usage{
.argsHash(formals, useLabels)
}
\arguments{
\item{formals}{named list with arguments}

\item{useLabels, }{logical, should be \code{TRUE} for now}
}
\value{
character with hash
}
\description{
This is copied code from madrat to create the hash for additional arguments
to `madrat::retrieveData` besides the regionmapping. This way, we can
recreate the file name of the tgz file with the input data.
}
\details{
Ideally, this should be an exported function of madrat rather than copied
code.
}
\author{
Robin Hasse
}
