% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.asNumeric}
\alias{.asNumeric}
\title{Convert to numeric vector}
\usage{
.asNumeric(x, warn = TRUE)
}
\arguments{
\item{x}{factor vector}

\item{warn}{logical, if \code{TRUE}, a warning is thrown when trying to
convert a non-factor}
}
\value{
numeric vector
}
\description{
Convert factor to numeric vector. This only works for factors with numeric
levels.
}
