% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateScenarioRuns.R
\name{.changeConfig}
\alias{.changeConfig}
\title{Change config switch}
\usage{
.changeConfig(configPath, param, value, comment = NULL)
}
\arguments{
\item{configPath}{character, file path to config}

\item{param}{character, name of switch}

\item{value}{value of switch}

\item{comment}{character, comment behind switch value}
}
\description{
Reset value of specified switch in given config. If the switch has not been
found in the config before, it is appended at the end to avoid disturbing the
nesting of the yaml file.
}
