% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.checkArmijoStep}
\alias{.checkArmijoStep}
\title{Check if the Armijo condition holds.
Return only data combinations for which it does not hold.}
\usage{
.checkArmijoStep(
  prevStep,
  stepSizeParams,
  outerObjective,
  sensitivityArmijo,
  varName = "fA",
  factorMin = 1
)
}
\arguments{
\item{prevStep}{data frame with data combinations that did not satisfy the Armijo condition in the previous step}

\item{stepSizeParams}{data frame with step size and related parameters}

\item{outerObjective}{data frame containing the value of the outer objective function.
Needs to contain the column \code{f} and the column specified by \code{varName}.}

\item{sensitivityArmijo}{numeric, parameter of the Armijo condition specifying how strict the condition is}

\item{varName}{character, outer objective variable to test the Armijo condition on}

\item{factorMin}{numeric, additional factor to modify the Armijo condition}
}
\description{
Check if the Armijo condition holds.
Return only data combinations for which it does not hold.
}
