% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.checkHeuristicStep}
\alias{.checkHeuristicStep}
\title{Check if the condition of the heuristic step size adaptation holds:
Check if the outer objective does not increase.
Return only data combinations for which the condition does not hold, i.e. the outer objective increases}
\usage{
.checkHeuristicStep(prevStep, outerObjective)
}
\arguments{
\item{prevStep}{data frame with data combinations that did not satisfy the heuristic condition in the previous step}

\item{outerObjective}{data frame containing the value of the outer objective function.
Needs to contain the columns \code{f} and \code{fA}.}
}
\description{
Check if the condition of the heuristic step size adaptation holds:
Check if the outer objective does not increase.
Return only data combinations for which the condition does not hold, i.e. the outer objective increases
}
