% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.checkStepCondition}
\alias{.checkStepCondition}
\title{Check if the Armijo condition holds or if a local minimum of the outer objective has been identified.
Return only data combinations for which it does not hold.}
\usage{
.checkStepCondition(
  prevStep,
  stepSizeParams,
  outerObjective,
  sensitivityArmijo
)
}
\arguments{
\item{prevStep}{data frame with data combinations that did not satisfy the step size condition in the previous step}

\item{stepSizeParams}{data frame with step size and related parameters}

\item{outerObjective}{data frame containing the value of the outer objective function.
Needs to contain the column \code{f} and the column specified by \code{varName}.}

\item{sensitivityArmijo}{numeric, parameter of the Armijo condition specifying how strict the condition is}
}
\description{
Check if the Armijo condition holds or if a local minimum of the outer objective has been identified.
Return only data combinations for which it does not hold.
}
