% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.checkStoppingCriterion}
\alias{.checkStoppingCriterion}
\title{Check for the stopping criterion of the iteration}
\usage{
.checkStoppingCriterion(
  outerObjective,
  p_calibTarget,
  threshold,
  zeroFlow = FALSE
)
}
\arguments{
\item{outerObjective}{data frame containing the value of the outer objective function.}

\item{p_calibTarget}{list of data frames with historic flows}

\item{threshold}{list of numeric, absolute and relative threshold of the stopping criterion}

\item{zeroFlow}{logical, whether renovation flow data should be filtered for zero flows}
}
\description{
Returns \code{TRUE} if and only if both the absolute and the relative stopping criterion
are satisfied for all subsets.
}
