% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.combineOuterObjective}
\alias{.combineOuterObjective}
\title{Read in brick results and compute the total outer objective function
by combining resuls from construction and renovation.}
\usage{
.combineOuterObjective(
  m,
  outerObjective,
  p_constructionCalibTarget,
  p_renovationCalibTarget,
  tcalib,
  varName = "f"
)
}
\arguments{
\item{m}{Gams transfer container to read brick results from}

\item{outerObjective}{data frame to write the outer objective to}

\item{p_constructionCalibTarget}{data frame with historic construction flows}

\item{p_renovationCalibTarget}{data frame with historic renovation flows}

\item{tcalib}{numeric, calibration time steps}

\item{varName}{character, column name in \code{outerObjective} to write the result to.
Should be one of 'f', 'fA', 'fMin'.}
}
\description{
Read in brick results and compute the total outer objective function
by combining resuls from construction and renovation.
}
