% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.computeDescentDirection}
\alias{.computeDescentDirection}
\title{Compute the descent direction from Brick results}
\usage{
.computeDescentDirection(
  m,
  dims,
  tcalib,
  flow = c("construction", "renovation", "renovationBS", "renovationHS")
)
}
\arguments{
\item{m}{Gams transfer container with current Brick results}

\item{dims}{character, dimensions of the desired result object}

\item{tcalib}{numeric, time steps to calibrate on}

\item{flow}{character, either 'construction' or 'renovation'}
}
\description{
Read in objective function values for slightly shifted optimization variables
and compute the gradient of the objective function.
Set the adjustment term \code{d} to the negative of the gradient.
}
