% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.computeOuterObjective}
\alias{.computeOuterObjective}
\title{Evaluate the outer objective function.}
\usage{
.computeOuterObjective(m, gamsVar, target, dims, tcalib, agg = NULL)
}
\arguments{
\item{m}{Gams transfer container object to read set values from}

\item{gamsVar}{data frame with brick results for the given variable}

\item{target}{data frame with historic data}

\item{dims}{character, dimensions of data}

\item{tcalib}{numeric, calibration time steps}

\item{agg}{character, dimension(s) to aggregate the data over}
}
\description{
In case of the logit calibration, this outer objective function is virtual:
We are not actually performing an optimization algorithm on this function,
but only use it for the step size adaptation.
}
