% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.computeOuterObjectiveStock}
\alias{.computeOuterObjectiveStock}
\title{Evaluate the outer objective function for the stock}
\usage{
.computeOuterObjectiveStock(
  m,
  outerObjective,
  p_stockCalibTarget,
  tcalib,
  varName = "f"
)
}
\arguments{
\item{m}{Gams transfer object to read the stock variable from}

\item{outerObjective}{data frame to write the outer objective to}

\item{p_stockCalibTarget}{data frame with historic stock}

\item{tcalib}{numeric, calibration time steps}

\item{varName}{character, column name in \code{outerObjective} to write the result to.
Should be one of 'f', 'fA', 'fMin'.}
}
\description{
Evaluate the outer objective function for the stock
}
