% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyHistoryGdx.R
\name{.cropParamsToThist}
\alias{.cropParamsToThist}
\title{crop gdx parameters to historic periods}
\usage{
.cropParamsToThist(gdx, thist)
}
\arguments{
\item{gdx}{character, file path to gdx}

\item{thist}{numeric vector of historic periods}
}
\description{
remove all records of temporal parameters that are outside of \code{thist}
and make zero values explicit filling with EPS.
}
