% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.filter}
\alias{.filter}
\title{Filter data frame rows with reference data frames}
\usage{
.filter(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{data frames that only have columns existing also in \code{x}}
}
\value{
filtered data frame
}
\description{
Performs successive semi_joins with passed data frames by all common columns.
Useful to filter allowed renovations or vintages that exist.
}
\author{
Robin Hasse
}
