% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParameters.R
\name{.filterLevel}
\alias{.filterLevel}
\title{filter rows with specified entry in column}
\usage{
.filterLevel(df, lvl, switchName = "", lvlCol = "level")
}
\arguments{
\item{df}{data.frame}

\item{lvl}{value used to select rows}

\item{switchName}{character, corresponding BRICK switch name (only used for
more informative error message)}

\item{lvlCol}{character, colname containing \code{lvl}}
}
\value{
data.frame with selected rows without \code{lvlCol}
}
\description{
used to select a specifc level or scenario from a data frame with alternative
values.
}
