% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConfig.R
\name{.findCfg}
\alias{.findCfg}
\title{Find config file path}
\usage{
.findCfg(config, configFolder, isFinalCfg)
}
\arguments{
\item{config}{character, config file, either a path to a yaml file or the
name of the file in `inst/config/`}

\item{configFolder}{character, directory to search for configs. If NULL, the
BRICK-internal config folder is used.}

\item{isFinalCfg}{logical, is this the final config and not an intermediate?}
}
\value{
file path to config
}
\description{
Search for config file in multiple steps. The file is found if
  * config is a full file path already
  * config is the name of a file in the config folder
  * there is exactly one file in \code{configFolder} matching the pattern
    passed via config
}
