% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.findMinimumStepSize}
\alias{.findMinimumStepSize}
\title{Determine the so far minimum step size.
If the outerObjective is increasing again after a decrease,
set the step size to the value yielding the current minimum.}
\usage{
.findMinimumStepSize(totalStep, stepSizeParams, outerObjective)
}
\arguments{
\item{totalStep}{data frame with combinations to perform the adjustment ot the step size on}

\item{stepSizeParams}{data frame with the parameters of the step size adjustment algorithm}

\item{outerObjective}{data frame containing the value of the outer objective function.}
}
\description{
Determine the so far minimum step size.
If the outerObjective is increasing again after a decrease,
set the step size to the value yielding the current minimum.
}
