% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyHistoryGdx.R
\name{.findOriginGdxFile}
\alias{.findOriginGdxFile}
\title{find origin gdx file}
\usage{
.findOriginGdxFile(originGdx)
}
\arguments{
\item{originGdx}{character, file path to run or gdx file used as historical}
}
\value{
file path to origin gdx file
}
\description{
find origin gdx file
}
