% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.findRegionMapping}
\alias{.findRegionMapping}
\title{Find region mapping}
\usage{
.findRegionMapping(regionmapping)
}
\arguments{
\item{regionmapping}{character. If \code{NULL}, country regions are used. It
can otherwise be a path to a mapping file (length one) or a vector of
length two with the file name and the location (\code{where}) of a madrat
regional mapping.}
}
\value{
data frame with region mapping
}
\description{
Find region mapping
}
