% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParameters.R
\name{.freezeParameters}
\alias{.freezeParameters}
\title{Freeze Parameters}
\usage{
.freezeParameters(m, params, period)
}
\arguments{
\item{m}{gamstransfer Container with parameters}

\item{params}{character vector of parameter names}

\item{period}{numeric, period that all values are freezed to; has to be an
element of ttot}
}
\value{
gams Container
}
\description{
Select parameters and freeze them to the value in a selected period
}
\author{
Robin Hasse
}
