% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.getBundleRunPaths}
\alias{.getBundleRunPaths}
\title{Get run paths from settings file}
\usage{
.getBundleRunPaths(settings)
}
\arguments{
\item{settings}{named list with bundle settings}
}
\value{
named vector of run paths
}
\description{
Get run paths from settings file
}
