% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateScenarioRuns.R
\name{.getConfigs}
\alias{.getConfigs}
\title{Get configs}
\usage{
.getConfigs(settings)
}
\arguments{
\item{settings}{named list with bundle settings}
}
\value{
named list of selected config files
}
\description{
Get configs
}
