% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.getFileName}
\alias{.getFileName}
\title{Get file name without ending}
\usage{
.getFileName(filePath)
}
\arguments{
\item{filePath}{character vector, file path including file ending}
}
\value{
file name without ending and directory
}
\description{
Get file name without ending
}
