% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDimMap.R
\name{.getMapping}
\alias{.getMapping}
\title{Get BRICK mapping}
\usage{
.getMapping(mapping)
}
\arguments{
\item{mapping}{character or list of characters, name of BRICK mapping. If
\code{"dim"} or \code{"agg"} is passed, all corresponding dimension
mappings are returned as named list.}
}
\value{
data frame or list of data frames
}
\description{
Get BRICK mapping
}
\author{
Robin Hasse
}
