% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateScenarioRuns.R
\name{.getPlaceholderConfigNames}
\alias{.getPlaceholderConfigNames}
\title{Get placeholder config names}
\usage{
.getPlaceholderConfigNames()
}
\value{
vector of config names
}
\description{
Get placeholder config names
}
