% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateScenarioRuns.R
\name{.getSettingsPath}
\alias{.getSettingsPath}
\title{Get path from settings file}
\usage{
.getSettingsPath(settings, pathName)
}
\arguments{
\item{settings}{named list with bundle settings}

\item{pathName}{character, name of path in settings file}
}
\value{
absolute path
}
\description{
Relative paths are assumed to relate to the directory of the settings file.
}
\author{
Robin Hasse
}
