% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.identicalPath}
\alias{.identicalPath}
\title{Test paths for equality}
\usage{
.identicalPath(x, y)
}
\arguments{
\item{x, y}{character, paths}
}
\value{
logical indicating if paths are identical
}
\description{
Test paths for equality
}
