% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateScenarioRuns.R
\name{.initialiseScenarios}
\alias{.initialiseScenarios}
\title{Initialise scenarios}
\usage{
.initialiseScenarios(settings, bundleSettings)
}
\arguments{
\item{settings}{named list with bundle settings}

\item{bundleSettings}{character, path to settings file. If the file doesn't
exist, a new file is written as a template.}
}
\value{
updated settings
}
\description{
Initialise scenarios
}
\author{
Robin Hasse
}
