% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startModel.R
\name{.initNextScens}
\alias{.initNextScens}
\title{Initialise next scenarios}
\usage{
.initNextScens(path)
}
\arguments{
\item{path}{character, path to the historic run folder}
}
\description{
Initialise the runs that start from a given run as historic run
}
