% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.initOptimVarCon}
\alias{.initOptimVarCon}
\title{Initialize the data frame for the optimization variable of construction}
\usage{
.initOptimVarCon(mInput, tcalib)
}
\arguments{
\item{mInput}{Gamstransfer container with the input data}

\item{tcalib}{numeric, calibration time periods}
}
\description{
Initialize the data frame for the optimization variable of construction
}
