% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.initOptimVarRen}
\alias{.initOptimVarRen}
\title{Initialize the data frame for the optimization variable of renovation}
\usage{
.initOptimVarRen(mInput, tcalib, renAllowed)
}
\arguments{
\item{mInput}{Gamstransfer container with the input data}

\item{tcalib}{numeric, calibration time periods}

\item{renAllowed}{data frame with allowed renovation transitions}
}
\description{
Initialize the data frame for the optimization variable of renovation
}
