% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.initOuterObjective}
\alias{.initOuterObjective}
\title{Initialize the data frame for the objective function value}
\usage{
.initOuterObjective(mInput, tcalib)
}
\arguments{
\item{mInput}{Gamstransfer container with the input data}

\item{tcalib}{numeric, calibration time periods}
}
\description{
Initialize the data frame for the objective function value
}
