% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.initStepSize}
\alias{.initStepSize}
\title{Compute the initial step size of the step size iteration algorithm}
\usage{
.initStepSize(i, stepSizeParams, outerObjective, stepSizeInit)
}
\arguments{
\item{i}{numeric, iteration number of overall calibration procedure}

\item{stepSizeParams}{data frame with parameters for the step size adaptation}

\item{outerObjective}{data frame with previous and current values of outer objective function;
needs to contain the columns \code{fPrev} and \code{f}}

\item{stepSizeInit}{numeric, lower bound for the initial step size}
}
\value{
data frame with the initial step size and initial values for the minimum outer objective
  and corresponding step size
}
\description{
Compute the initial step size of the step size iteration algorithm
}
