% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.insertBetween}
\alias{.insertBetween}
\title{Insert element between each element of a vector}
\usage{
.insertBetween(x, insert)
}
\arguments{
\item{x}{vector}

\item{insert}{element to insert}
}
\value{
vector of length \code{length(x) * 2 - 1}
}
\description{
Insert element between each element of a vector
}
\author{
Robin Hasse
}
