% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.isCascade}
\alias{.isCascade}
\title{Check if an object is a cascade of BRICK configs}
\usage{
.isCascade(x)
}
\arguments{
\item{x}{any R object (should be a named list if you expect a cascade)}
}
\value{
logical
}
\description{
Check if an object is a cascade of BRICK configs
}
