% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.isConfig}
\alias{.isConfig}
\title{Check if an object is a BRICK config}
\usage{
.isConfig(x)
}
\arguments{
\item{x}{any R object (should be a named list if you expect a config)}
}
\value{
logical
}
\description{
Check if an object is a BRICK config
}
