% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.isTRUE}
\alias{.isTRUE}
\title{Test if vector elements are TRUE}
\usage{
.isTRUE(x)
}
\arguments{
\item{x}{vector}
}
\value{
logical vector
}
\description{
Perform \code{isTRUE} on each vecotr element
}
\author{
Robin Hasse
}
