% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.logLikelihood}
\alias{.logLikelihood}
\title{Calculate log-Likelihood of historical values}
\usage{
.logLikelihood(res, target)
}
\arguments{
\item{res}{numeric, result values}

\item{target}{numeric, historic values}
}
\value{
numeric, negative of the log likelihood
}
\description{
Calculate log-Likelihood of historical values
}
