% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createParameters.R
\name{.makeIdentVin}
\alias{.makeIdentVin}
\title{Make vintage characteristic identical}
\usage{
.makeIdentVin(x)
}
\arguments{
\item{x}{data.frame with brick parameter, requires vin and value column}
}
\value{
data.frame with identical structure but averaged values
}
\description{
fill value column with average across vintages
}
