% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.namedLapply}
\alias{.namedLapply}
\title{Lapply with output named by given list}
\usage{
.namedLapply(nm, ...)
}
\arguments{
\item{nm}{list of characters, lapply input list and name of output}

\item{...}{further inputs to lapply}
}
\value{
named list
}
\description{
Lapply with output named by given list
}
