% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.normaliseMatchingPath}
\alias{.normaliseMatchingPath}
\title{Normalise matching path}
\usage{
.normaliseMatchingPath(matchingRun)
}
\arguments{
\item{matchingRun}{character either a subfolder in the brick matching madrat
source or an existing run path}
}
\value{
path to a matching run
}
\description{
Normalise matching path
}
