% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{.onLoad}
\alias{.onLoad}
\title{Check if gamstransfer is available}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{not used}

\item{pkgname}{not used}
}
\description{
Just a check to get a helpful error if gamstransfer is missing. Once the
package is available on GitHub, it should be listed as a dependency and this
file can be removed. The check is disabled during contiunous integration (CI)
i.a. GitHub actions and during \code{devtools:check} which is called by
\code{lucode2::buildLibrary}.
}
\author{
Robin Hasse
}
