% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConfig.R
\name{.overwriteList}
\alias{.overwriteList}
\title{Overwrite list with another list}
\usage{
.overwriteList(x, y, isFinalCfg, defaultCfgPath)
}
\arguments{
\item{x}{named list, provides the structure and default values that can be
overwritten by \code{y}.}

\item{y}{named list, overwrites \code{x} wherever it has values different
from NULL. Cannot have keys that are not specified in \code{x}.}

\item{isFinalCfg}{logical, is this the final config and not an intermediate?}

\item{defaultCfgPath}{character, path to default config. Only used for more
helpful error message.}
}
\value{
named list with the structure of \code{x} that is (partly)
  overwritten by \code{y}.
}
\description{
Overwrite a named list with another named list. The result corresponds to the
overwritten list unless a value is overwritten by the overwriting list.
}
\details{
This function is called recursively until the default config is reached.
Therefor, the list structure will always correspond to the default config.
No config based directly or indirectly on the default can have list keys that
the default config doesn't have.
}
