% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.printConvergenceMessage}
\alias{.printConvergenceMessage}
\title{Print message that the calibration converged.}
\usage{
.printConvergenceMessage(iteration, threshold)
}
\arguments{
\item{iteration}{numeric, number of current iteration}

\item{threshold}{list of numeric, absolute and relative threshold of the stopping criterion}
}
\description{
Print message that the calibration converged.
}
