% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.readCalibTarget}
\alias{.readCalibTarget}
\title{Read calibration targets from input folder}
\usage{
.readCalibTarget(dims, mInput)
}
\arguments{
\item{dims}{named character, dimensions of each variable of the calibration}

\item{mInput}{gamstransfer container with the input.gdx}
}
\description{
Read calibration targets from input folder
}
