% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConfig.R
\name{.readCfg}
\alias{.readCfg}
\title{Read config file}
\usage{
.readCfg(file, warn, commentsAsAttr = FALSE)
}
\arguments{
\item{file}{character, path to config file}

\item{warn}{logical, raise a warning if a title has been declared in the
config}

\item{commentsAsAttr}{logical, if \code{TRUE}, commented lines are read as
yaml of config attributes}
}
\value{
named list with config parameters
}
\description{
Read yaml file from given path, fill title with file name and save the file
path as an attribute.
}
