% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateScenarioRuns.R
\name{.readConfigSwitch}
\alias{.readConfigSwitch}
\title{Read config Switch}
\usage{
.readConfigSwitch(configPath, switch)
}
\arguments{
\item{configPath}{character, path to config file}

\item{switch}{character, name of config switch}
}
\value{
named list with switch value and comment. \code{NULL} is switch is
  missing in config
}
\description{
Read config Switch
}
