% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startModel.R
\name{.readInitArgs}
\alias{.readInitArgs}
\title{Read model initialisation arguments}
\usage{
.readInitArgs(path)
}
\arguments{
\item{path}{character, path to run folder}
}
\value{
named list of arguments to \code{initModel}
}
\description{
Since its function call, \code{initModel} might have changed arguments,
namely the restart options.
}
