% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateScenarioRuns.R
\name{.readSettings}
\alias{.readSettings}
\title{Read bundle settings}
\usage{
.readSettings(path)
}
\arguments{
\item{path}{character, path to bundle settings file}
}
\value{
named list with settings
}
\description{
This function also checks that the settings contain all required keys and
only arguments to \code{initModel} as further arguments.
}
