% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.refMapName}
\alias{.refMapName}
\title{Create reference mapping name}
\usage{
.refMapName(ref, file = FALSE)
}
\arguments{
\item{ref}{character, reference name}

\item{file}{file extension appended unless set to \code{FALSE}}
}
\value{
character with reference mapping name
}
\description{
Create reference mapping name
}
\author{
Robin Hasse
}
