% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.rejectErrorRun}
\alias{.rejectErrorRun}
\title{Reject step sizes with unsuccesful Gams run}
\usage{
.rejectErrorRun(
  path,
  outerObjective,
  gamsSuccess,
  iteration,
  iterationStepSize,
  varName = "fA",
  gdxName = "calibrationA.gdx"
)
}
\arguments{
\item{path}{character, run path where all files, in particular Gams files, are stored}

\item{outerObjective}{data frame with data on outer objective}

\item{gamsSuccess}{data frame containing whether Gams was successful for each subset}

\item{iteration}{numeric, current calibration iteration}

\item{iterationStepSize}{numeric, current iteration of step size adaptation}

\item{varName}{character, column of outer objective to be modified}

\item{gdxName}{character, name of the gdx file to be copied.}
}
\value{
data frame, outer objective with modifications in the column given by \code{varName}
}
\description{
If Gams was not successful, set the outer objective to infinity
so that it cannot be accepted by the step size adaptation.
Copy relevant files to trace the error (gdx, main.log and main.lst)
}
