% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.setSwitch}
\alias{.setSwitch}
\title{Set config switch}
\usage{
.setSwitch(config, ...)
}
\arguments{
\item{config}{named list, config or cascade of configs}

\item{...}{named arguments wit switch values}
}
\value{
config or cascade with changed switch value
}
\description{
This function preserves attributes that usually get lost when doing simple
replacements.
}
