% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.sumSquare}
\alias{.sumSquare}
\title{Calculate sum of squared differences between results and historic values}
\usage{
.sumSquare(res, target)
}
\arguments{
\item{res}{numeric, result values}

\item{target}{numeric, historic values}
}
\description{
Calculate sum of squared differences between results and historic values
}
