% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.unique}
\alias{.unique}
\title{Extract elements that are unique to gams}
\usage{
.unique(x)
}
\arguments{
\item{x}{a vector or a data frame}
}
\value{
an object of the same kind but possibly fewer elements. Vectors can
  be shorter and data frames can have fewer rows
}
\description{
Wrapper around \code{unique} that also drops elements that differ in cases
but are identical to gams as it is case-insensitive.
}
\author{
Robin Hasse
}
