% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startModel.R
\name{.updateBundleSettings}
\alias{.updateBundleSettings}
\title{Update bundle settings}
\usage{
.updateBundleSettings(path, title)
}
\arguments{
\item{path}{character, path to the historic run folder}

\item{title}{character, path to the historic run folder}
}
\description{
Write run path into settings file
}
