% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateScenarioRuns.R
\name{.updateRefScens}
\alias{.updateRefScens}
\title{Update referenced scenario}
\usage{
.updateRefScens(configs, settings, param, allowSelfRef = TRUE)
}
\arguments{
\item{configs}{named character vector of config file paths}

\item{settings}{named list with bundle settings}

\item{param}{character, name of config switch}

\item{allowSelfRef}{logical, if TRUE, the config parameter can reference
itself}
}
\value{
update settings
}
\description{
Update referenced scenario
}
