% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.updateX}
\alias{.updateX}
\title{Update the optimization variable 'x'}
\usage{
.updateX(optimVar, deviation, stepSizeParams, dims, nameTo = "x")
}
\arguments{
\item{optimVar}{data frame with the optimization variable 'x' and optionally and 'xA'}

\item{deviation}{data frame with deviation and adjustment term 'd'}

\item{stepSizeParams}{data frame with parameters of step size adaptation, including step size 'stepSize'}

\item{dims}{character, dimensions of the optimization variable}

\item{nameTo}{character, optimization variable to write the result to}
}
\description{
Adjust 'x' by adding the adjustment term 'd' multiplied by the step size
}
