% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.updateXSelect}
\alias{.updateXSelect}
\title{Update the optimization variable 'x' for selected combinations only}
\usage{
.updateXSelect(armijoStep, optimVar, deviation, stepSizeParams, dims)
}
\arguments{
\item{armijoStep}{data frame with combinations to perform the adjustment of 'x' on}

\item{optimVar}{data frame with the optimization variables}

\item{deviation}{data frame with the deviation from historic data and the adjustment parameter 'd'}

\item{stepSizeParams}{data frame with the parameters of the step size adaptatation procedure}

\item{dims}{character, dimensions of the optimization variable}
}
\description{
Update the optimization variable 'x' for selected combinations only
}
