% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.writeCostIntang}
\alias{.writeCostIntang}
\title{Write the intangible costs to a .csv file}
\usage{
.writeCostIntang(file, optimVar, xinit, dims, tcalib)
}
\arguments{
\item{file}{character, path to file to which results should be written}

\item{optimVar}{data frame with optimization variable, needs to contain column 'x'}

\item{xinit}{data frame with initial intangible specific costs}

\item{dims}{character, dimensions of the optimization variable}

\item{tcalib}{numeric, calibration time steps}
}
\description{
Write the intangible costs to a .csv file
}
