% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.writeShareRen}
\alias{.writeShareRen}
\title{Write the renovation share of heating systems to a csv file}
\usage{
.writeShareRen(m, path)
}
\arguments{
\item{m}{Gams container with the calibration output.gdx}

\item{path}{character, path to the calibration run}
}
\description{
Write the renovation share of heating systems to a csv file
}
