% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.writeStockRes}
\alias{.writeStockRes}
\title{Write the stock of the calibration result to a csv file}
\usage{
.writeStockRes(m, path)
}
\arguments{
\item{m}{Gams container with the calibration output.gdx}

\item{path}{character, path to the calibration run}
}
\description{
Write the stock of the calibration result to a csv file
}
