% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandSets.R
\name{expandSets}
\alias{expandSets}
\title{Expand set values to data frame}
\usage{
expandSets(..., .m = NULL)
}
\arguments{
\item{...}{gams sets}

\item{.m}{gams Container with sets referenced in \code{...}.}
}
\value{
data.frame with full crossing of set entries
}
\description{
Create a data frame that has a column for each set and the full crossing of
all set entries as rows
}
\author{
Robin Hasse
}
