% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolateShareRen.R
\name{extrapolateShareRen}
\alias{extrapolateShareRen}
\title{Extrapolate share of initial heating system to replace}
\usage{
extrapolateShareRen(path)
}
\arguments{
\item{path}{character, directory of matching folder}
}
\value{
data frame with \code{v_shareRenHSinit} extrapolated to future time
  steps
}
\description{
Extrapolation to future time steps such that the matching results can be used
for temporal scopes that go beyond the matching scope.
}
\author{
Robin Hasse
}
