% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGdxFile.R
\name{findGdxFile}
\alias{findGdxFile}
\title{Find GDX file}
\usage{
findGdxFile(gdx, outputFolder = NULL, gdxNames = "output.gdx")
}
\arguments{
\item{gdx}{character, either a file path, a run directory or the name of a
run in the output folder}

\item{outputFolder}{character, path to output folder. This argument is
required if only a run name is passed as \code{gdx}.}

\item{gdxNames}{character, file names to consider when looking for GDX files
in run folders. If you allow multiple file names, they are tested in the
given order and the first match is returned (start with highest priority).}
}
\value{
path to gdx file
}
\description{
Find file path to the most explicitly referenced GDX file
}
\author{
Robin Hasse
}
