% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLastRun.R
\name{findLastRun}
\alias{findLastRun}
\title{Determine the latest path of a run}
\usage{
findLastRun(outputFolder)
}
\arguments{
\item{outputFolder}{character, output folder to search in}
}
\description{
Search the given output folder for the run with the most recent time stamp
and return the path to this run.
If the output folder contains only one run, return its path, also if it does not contain a time stamp.
}
\author{
Ricarda Rosemann
}
