% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBrickMapping.R
\name{getBrickMapping}
\alias{getBrickMapping}
\title{Retrieve mapping file from BRICK}
\usage{
getBrickMapping(
  name,
  type = "sectoral",
  error.missing = TRUE,
  returnPathOnly = FALSE
)
}
\arguments{
\item{name}{character, file name of the mapping file}

\item{type}{character, Mapping type (e.g. "regional", or "sectoral")}

\item{error.missing}{logical, return error if the file does not exist}

\item{returnPathOnly}{logical, file name of the mapping file}
}
\description{
Retrieve mapping file from BRICK
}
\author{
Robin Hasse
}
