% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDimMap.R
\name{getDimMap}
\alias{getDimMap}
\title{Get Brick dimension mapping}
\usage{
getDimMap(dim, granularity = NULL)
}
\arguments{
\item{dim}{character, dimension name}

\item{granularity}{character, name of BRICK granularity}
}
\value{
data.frame with dimension mapping
}
\description{
Get Brick dimension mapping
}
\author{
Robin Hasse
}
