% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertGamsCode.R
\name{insertGamsCode}
\alias{insertGamsCode}
\title{Insert code into gams scripts}
\usage{
insertGamsCode(path, script, section, insert, keepComments = TRUE)
}
\arguments{
\item{path}{character, description}

\item{script}{characer, name of gams script}

\item{section}{character, section name}

\item{insert}{character vector, code line that are inserted}

\item{keepComments}{logical, if TRUE, lines starting with \code{!!} are kept
and the inserted lines follow below}
}
\description{
Lines between marker comments are replaced by given lines. The respective
script file gets overwritten.
}
\details{
Before
\preformatted{
  !! AUTOCODE.<section>
  <old code>
  !! AUTOCODE_end.<section>
}

After
\preformatted{
  !! AUTOCODE.<section>
  <insert>
  !! AUTOCODE_end.<section>
}
}
\author{
Robin Hasse
}
