% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMatchingCode.R
\name{insertMatchingCode}
\alias{insertMatchingCode}
\title{Insert reference-dependent code for matching into gams scripts}
\usage{
insertMatchingCode(path)
}
\arguments{
\item{path}{character vector with folders to run the model in}
}
\description{
Insert reference-dependent code for matching into gams scripts
}
\author{
Robin Hasse
}
