% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSlurmAvailable.R
\name{isSlurmAvailable}
\alias{isSlurmAvailable}
\title{Is Slurm Avaiable}
\usage{
isSlurmAvailable()
}
\value{
\code{logical(1)}. Whether slurm is available.
}
\description{
Checks whether slurm is available so that jobs can be submitted, e.g., via \code{sbatch ...}.
}
