% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToDf.R
\name{listToDf}
\alias{listToDf}
\title{Convert named list to data.frame}
\usage{
listToDf(x, split = NULL)
}
\arguments{
\item{x}{named list with keys as column names}

\item{split}{character, split names by this pattern and assume same value for
each separated item}
}
\value{
data.frame with column for each key
}
\description{
Convert named list to data.frame
}
\author{
Robin Hasse
}
