% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToGdx.R
\name{listToGdx}
\alias{listToGdx}
\title{Prepare data frame for GDX}
\usage{
listToGdx(setList)
}
\arguments{
\item{setList}{List of character vectors set names as names and their
respective elements as character vector}
}
\value{
A list with a particular structure required by [gdxrrw::wgdx()].
}
\description{
Create an object that can be handled by [gdxrrw::wgdx()] to write the data
frame as a gams set into a gdx file.
}
\details{
This function takes an array and creates a list ready to be written into a
gdx file. It will currently always create an output for a gams parameter.
}
\author{
Robin Hasse
}
