% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCalibrationTarget.R
\name{loadCalibrationTarget}
\alias{loadCalibrationTarget}
\title{Load calibration targets from matching results}
\usage{
loadCalibrationTarget(config, inputDir, path)
}
\arguments{
\item{config}{named list with run configuration}

\item{inputDir}{directory of input folder}

\item{path}{character, path to the run}
}
\description{
Load calibration targets from matching results
}
\author{
Ricarda Rosemann
}
