% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHandle.R
\name{makeHandle}
\alias{makeHandle}
\title{Named list to handle string}
\usage{
makeHandle(lst, type = c("gams", "model"))
}
\arguments{
\item{lst}{named list of flags}

\item{type}{character, type of flags (either `"gams"` or `"model`)}
}
\value{
character of flags
}
\description{
Helper function to turn a named list into a string appended to the gams
command line call
}
\details{
Gams parameters are lower case flags and follow one minus `-`. Model switches
are upper case flags and follow two minuses `--`.
}
\author{
Robin Hasse
}
