% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyCosts.R
\name{modifyCosts}
\alias{modifyCosts}
\title{Modify costs as specified in config}
\usage{
modifyCosts(costs, costMod, t0)
}
\arguments{
\item{costs}{data frame, containing costs values}

\item{costMod}{named list, specification on how to modify costs}

\item{t0}{numeric, initial time period}
}
\description{
Modify costs as specified in config
}
\author{
Ricarda Rosemann
}
