% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRefDeviation.R
\name{plotRefDeviation}
\alias{plotRefDeviation}
\title{Plot heat map of reference deviation}
\usage{
plotRefDeviation(path, dropZeroWeightRefs = TRUE)
}
\arguments{
\item{path}{character, directory of output folder}

\item{dropZeroWeightRefs}{boolean, if TRUE, references with zero weight are
not shown}
}
\description{
Plot heat map of reference deviation
}
\author{
Robin Hasse
}
