% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSankey.R
\name{plotSankey}
\alias{plotSankey}
\title{Plot Sakey diagram}
\usage{
plotSankey(path, vin = NULL)
}
\arguments{
\item{path}{character, path to the run}
}
\description{
Plot Sakey diagram
}
\author{
Robin Hasse
}
