% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummary.R
\name{plotSummary}
\alias{plotSummary}
\title{Plot Summary of a run}
\usage{
plotSummary(path, showHistStock = FALSE)
}
\arguments{
\item{path}{character, path to the run}

\item{showHistStock}{logical, show given historic next to the modeled stock}
}
\description{
Plot an overview of the stock and flows
}
\author{
Robin Hasse
}
