% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareFuelPrices.R
\name{prepareFuelPrices}
\alias{prepareFuelPrices}
\title{Prepare fuel prices}
\usage{
prepareFuelPrices(carbonPrice = NULL, regions, periods, result = "fuelPrices")
}
\arguments{
\item{carbonPrice}{data.frame with time series of carbon price}
}
\description{
Prepare fuel prices
}
\author{
Robin Hasse
}
