% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConfig.R
\name{readConfig}
\alias{readConfig}
\title{Read config file}
\usage{
readConfig(config = NULL, basisOf = NULL)
}
\description{
Read config file in yaml format
}
\details{
If no argument is given, the default config is used.
}
\author{
Robin Hasse
}
