% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInput.R
\name{readInput}
\alias{readInput}
\title{Read madrat input files from input folder}
\usage{
readInput(filename, dims, inputDir = NULL)
}
\arguments{
\item{filename}{character, filename}

\item{dims}{character vector, column names}

\item{inputDir}{character, directory of input folder}
}
\value{
data.frame with data of the given input file
}
\description{
Read madrat input files from input folder
}
\author{
Robin Hasse
}
