% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSymbol.R
\name{readSymbol}
\alias{readSymbol}
\title{Read symbol from gams container}
\usage{
readSymbol(m, symbol, selectArea = TRUE)
}
\arguments{
\item{m}{gams container}

\item{symbol}{character, name of gams object}

\item{selectArea}{logical, select area quantity and remove this dimension}
}
\description{
Read symbol from gams container
}
\author{
Robin Hasse
}
