% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMif.R
\name{reportMif}
\alias{reportMif}
\title{Create mif file for model run}
\usage{
reportMif(path, file = NULL, tmpl = NULL)
}
\arguments{
\item{path}{character, path to the run}

\item{file}{character, path of mif file. If \code{NULL}, default file name
in run folder will be used. If \code{FALSE}, no file is written and the mif
data is returned instead.}

\item{tmpl}{character, BRICK reporting template. There has to be a brickSets
mapping named with the same suffix: \code{brickSets_<tmpl>.yaml}. If NNULL,
the config setting is used.}
}
\description{
The mif file contains reporting variables for the given model run.
}
\author{
Robin Hasse
}
