% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restartModel.R
\name{restartModel}
\alias{restartModel}
\title{Start the model}
\usage{
restartModel(
  path = NULL,
  outputFolder = "output",
  copyScripts = FALSE,
  copyStartPoint = FALSE,
  recreateInputData = FALSE,
  copyHistory = FALSE,
  recreateMatchingData = FALSE,
  reaggregateMatching = FALSE,
  references = NULL
)
}
\arguments{
\item{path}{character vector with folders to run the model in}

\item{outputFolder}{directory of output folder}

\item{copyScripts}{boolean, copy gams scripts from source}

\item{copyStartPoint}{boolean, copy initial gdx}

\item{recreateInputData}{boolean, recreate input data}

\item{copyHistory}{boolean, copy history gdx}

\item{recreateMatchingData}{boolean, recreate matching data}

\item{reaggregateMatching}{boolean, reaggregate matching results}

\item{references}{named list, not implemented yet}
}
\description{
Run the model with given configuration.
}
\details{
This function creates a run folder with necessary gams files if missing. It
then computes the input data and finally runs the optimisation.
}
\author{
Robin Hasse
}
