% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reweightMatchingReferences.R
\name{reweightMatchingReferences}
\alias{reweightMatchingReferences}
\title{Update matching reference selection and weights}
\usage{
reweightMatchingReferences(path)
}
\arguments{
\item{path}{character, path to run}
}
\value{
gamstransfer container with updates reference gdx content
}
\description{
Read reference configuration from existing run und update the set of selected
references and the reference weights in the reference.gdx
}
\author{
Robin Hasse
}
