% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{runCalibration}
\alias{runCalibration}
\title{Run the brick calibration}
\usage{
runCalibration(
  path,
  parameters,
  tcalib,
  gamsOptions = NULL,
  switches = NULL,
  fileName = "main.gms",
  gamsCall = "gams"
)
}
\arguments{
\item{path}{character vector with folders to run gams in}

\item{parameters}{named list of calibration parameters}

\item{tcalib}{numeric, time periods to calibrate on}

\item{gamsOptions}{named list of GAMS options}

\item{switches}{named list of model switches}

\item{fileName}{character vector with gams file names}

\item{gamsCall}{system command to call gams}
}
\description{
Two methods for the Brick calibration are available: The logit calibration and the
optimization calibration.
}
\details{
The general procedure is as follows:
We aim to minimize the deviation between historic values and  Brick results.
An intangible component of the specific construction and renovation costs is iteratively
adjusted to achieve this. Thus the intangible costs serve as the optimization variable \code{x}
and a functional evaluation of the deviation represents the objective function \code{f}.

Brick is run iteratively and from each run, an adjustment term \code{d} is computed.
The optimization variable is adjusted according to \code{x = x + stepSize * d}, where
the step size is determined according to the Armijo step size adaptation algorithm.

The implementation is centered around the following objects:
\itemize{
\item The optimization variable \code{x} and its variant(s) required in the step size adaptation
are gathered in the data frame \code{optimVarCon} for construction and \code{optimVarRen}
for renovation.
\item The objective function value \code{f} and its variant(s) are gathered in the data
frame \code{outerObjective}
\item The adjustment term \code{d} and related variables if applicable are gathere in the
data frame \code{deviationCon} for construction and \code{deviationRen} for renovation.
\item The step size \code{stepSize} and related parameters requried for the step size adaptation
algorithm are gathered in the data frame \code{stepSizeParams}.
}
}
\author{
Ricarda Rosemann
}
