% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{runCalibrationLogit}
\alias{runCalibrationLogit}
\title{Call the logit brick calibration}
\usage{
runCalibrationLogit(
  path,
  parameters,
  tcalib,
  calibTarget,
  gamsOptions = NULL,
  switches = NULL,
  fileName = "main.gms",
  gamsCall = "gams"
)
}
\arguments{
\item{path}{character vector with folders to run gams in}

\item{parameters}{named list of calibration parameters}

\item{tcalib}{numeric, time periods to calibrate on}

\item{calibTarget}{list of data frames of calibration targets}

\item{gamsOptions}{named list of GAMS options}

\item{switches}{named list of model switches}

\item{fileName}{character vector with gams file names}

\item{gamsCall}{system command to call gams}
}
\description{
The adjustment term is computed as if Brick were a pure logit model:
\code{d = log(<brick result>/<historic value>)}.
}
\details{
Here the armijo step size algorithm may not be applicable; if this is the
case, the minimum objective over all armijo iterations is used.
}
