% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{runCalibrationOptim}
\alias{runCalibrationOptim}
\title{Call the optimization brick calibration}
\usage{
runCalibrationOptim(
  path,
  parameters,
  tcalib,
  dims,
  gamsOptions = NULL,
  switches = NULL,
  fileName = "main.gms",
  gamsCall = "gams"
)
}
\arguments{
\item{path}{character vector with folders to run gams in}

\item{parameters}{named list of calibration parameters}

\item{tcalib}{numeric, time periods to calibrate on}

\item{dims}{list of characters with dimensions of stock and flow variables}

\item{gamsOptions}{named list of GAMS options}

\item{switches}{named list of model switches}

\item{fileName}{character vector with gams file names}

\item{gamsCall}{system command to call gams}
}
\description{
This implements a gradient descent method to perform the minimization of the deviation.
The adjustment term is computed as the negative of the gradient of the objective function.
}
