% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGams.R
\name{runGams}
\alias{runGams}
\title{Run gams optimisation}
\usage{
runGams(
  path,
  gamsOptions = NULL,
  switches = NULL,
  fileName = "main.gms",
  gamsCall = "gams"
)
}
\arguments{
\item{path}{character vector with folders to run gams in}

\item{gamsOptions}{named list of GAMS options}

\item{switches}{named list of model switches}

\item{fileName}{character vector with gams file names}

\item{gamsCall}{system command to call gams}
}
\description{
Run the gams model in a given directory.
}
\details{
The file `input.gdx` has to exist in the given directory.
}
\author{
Robin Hasse
}
