% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivityRun.R
\name{sensitivityRun}
\alias{sensitivityRun}
\title{Modify costs as specified in config}
\usage{
sensitivityRun(
  config,
  method = "scale",
  allScales = NULL,
  hs = "gabo",
  fullRun = TRUE,
  ...
)
}
\arguments{
\item{config}{character, file name  of the config to be used}

\item{method}{character, specification on how to modify costs}

\item{allScales}{numeric, scale/values for cost modification}

\item{hs}{character, heating system for which the costs are modified}

\item{fullRun}{logical, whether to call brick;
can be switched off to evaluate results after conducting runs with SLURM}

\item{...}{named list, arguments passed to initModel}
}
\description{
Modify costs as specified in config
}
\author{
Ricarda Rosemann
}
