% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSlurmConfig.R
\name{setSlurmConfig}
\alias{setSlurmConfig}
\title{Set the SLURM configuration}
\usage{
setSlurmConfig(slurmQOS, tasks32, nmbrRegi)
}
\arguments{
\item{slurmQOS}{string, name of the desired QOS (Quality of Service)}

\item{tasks32}{boolean, specify whether a node with 32 tasks should be requested}

\item{numbrRegi}{numeric, number of regions}
}
\value{
string with SLURM configuration
}
\description{
Based on user input, construct the SLURM configuration string.
}
\details{
Check if SLURM configuration is admissible.
}
\author{
Ricarda Rosemann
}
