% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startModel.R
\name{startModel}
\alias{startModel}
\title{Start the model}
\usage{
startModel(config = NULL, path = NULL, outputFolder = NULL, references = NULL)
}
\arguments{
\item{config}{run configurations}

\item{path}{character vector with folders to run the model in}

\item{outputFolder}{directory of output folder}

\item{references}{named list of matching references}
}
\description{
Run the model with given configuration.
}
\details{
This function creates a run folder with necessary gams files if missing. It
then computes the input data and finally runs the optimisation.
}
\author{
Robin Hasse
}
