% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolMakeHandle.R
\name{toolMakeHandle}
\alias{toolMakeHandle}
\title{Named list to handle string}
\usage{
toolMakeHandle(lst, type = c("gams", "model"))
}
\description{
Helper function to turn a named list into a string appended to the gams
command line call
}
\details{
Gams parameters are lower case flags and follow one minus `-`. Model switches
are upper case flags and follow two minuses `--`.
}
\author{
Robin Hasse
}
