% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateMatching.R
\name{aggregateMatching}
\alias{aggregateMatching}
\title{Aggregate matching run results for calibration}
\usage{
aggregateMatching(path, config, overwrite = FALSE)
}
\arguments{
\item{path}{character, path to calibration run}

\item{config}{named list, configuration of calibration run}

\item{overwrite}{logical, should existing data be overwritten?}
}
\description{
Temporary function that prepares historic stock and flows for calibration
}
\details{
This functionality will migrate to mredgebuidlings but this infrastructure
requires more time to be developed.
}
\author{
Robin Hasse
}
