% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brick-package.R
\docType{package}
\name{brick-package}
\alias{brick}
\alias{brick-package}
\title{brick: Building sector model with heterogeuous renovation and construction of the stock}
\description{
This building stock model represents residential and commercial buildings at customisable regional and temporal resolution. The building stock is quantified in floor area and distinuished by building type (SFH/MFH) and location (rural/urban). In each building category, construction cohorts are tracked explicitly. This allows to characterise buildings specifically for each of subset of buildings. The evolution of the building stock follows from the flows of constructed, renovated and demolished buildings and is optimised under cost minimisation with a benefit for heterogeneity in the choice of construction and renovation alternatives. This benefit captures heterogeneity in the preferences of the aganets and the building structure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/brick}
}

}
\author{
\strong{Maintainer}: Robin Hasse \email{robin.hasse@pik-potsdam.de} (\href{https://orcid.org/0000-0003-1818-3186}{ORCID})

}
