% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGamsSuccess.R
\name{checkGamsSuccess}
\alias{checkGamsSuccess}
\title{Check whether Gams finished successfully}
\usage{
checkGamsSuccess(path, silent = FALSE)
}
\arguments{
\item{path}{character, path to search for gams output files}

\item{silent}{logical, whether no messages should be printed.}
}
\value{
logical, whether GAMS was successful for all subsets
}
\description{
Check which file is the gdx file written by Gams from main.log.
Read model and solver status from this gdx file and print.
}
\details{
Stop with error message if:
  - \code{main.log} does not exist
  - the gdx file does not exist
  - more than one gdx file was written.
}
\author{
Ricarda Rosemann
}
