% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyHistoryGdx.R
\name{copyHistoryGdx}
\alias{copyHistoryGdx}
\title{Copy history gdx to output folder}
\usage{
copyHistoryGdx(path, config, overwrite = FALSE, thistOnly = TRUE)
}
\arguments{
\item{path}{character vector with folders to write input data into}

\item{config}{named list with run configuration}

\item{overwrite}{logical, should existing input.gdx be overwritten?}

\item{thistOnly}{logical, crop temporal parameters to historic periods}
}
\description{
Copy history gdx to output folder
}
\author{
Robin Hasse
}
