% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyInitialGdx.R
\name{copyInitialGdx}
\alias{copyInitialGdx}
\title{Copy initial gdx to output folder}
\usage{
copyInitialGdx(path, config, overwrite = FALSE)
}
\arguments{
\item{path}{character vector with folders to write input data into}

\item{config}{named list with run configuration}

\item{overwrite}{logical, should existing input.gdx be overwritten?}
}
\description{
Copy initial gdx to output folder
}
\author{
Robin Hasse
}
