% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCalibrationTarget.R
\name{createCalibrationTarget}
\alias{createCalibrationTarget}
\title{Create calibration targets}
\usage{
createCalibrationTarget(path, outDir)
}
\arguments{
\item{path}{character, directory of matching folder}

\item{outDir}{character, directory to save files in}
}
\description{
read results from matching run and aggregate them spatially and temporally
to the resolution of calibration runs
}
\details{
This is a first implementation with many known issues/limitations:
\itemize{
 \item The temporal resolution is hard-coded.
 \item It is assumed that the matching ran on European countries and is to be
 aggregated to one EU27 region
 \item The temporal aggregation of renovation flows is non-trivial: It is
 generally possible that buildings can change states multiple time within one
 aggregated time step but they can't be tracked through renovations
 individually. Currently, we just assume the mean of renovations across all
 years of an aggregated time step and adjust the zero flows to match the
 stock. This procedure can lead to inconsistent and even negative flows. In
 fact, aggregated stock and flows are inconsistent in the current version.
}
}
\author{
Robin Hasse
}
