% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInputData.R
\name{createInputData}
\alias{createInputData}
\title{Create input data}
\usage{
createInputData(path, config, overwrite = FALSE)
}
\arguments{
\item{path}{character vector with folders to write input data into}

\item{config}{named list with run configuration}

\item{overwrite}{logical, should existing input.gdx be overwritten?}
}
\description{
Create a complete set of input data for the gams optimisation.
}
\details{
This function reads static input data according to the input data revision in
the config and creates all required sets and parameters for the gams
optimisation depending on the switches in the config.
}
\author{
Robin Hasse
}
