% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMatchingData.R
\name{createMatchingData}
\alias{createMatchingData}
\title{Create data for reference matching}
\usage{
createMatchingData(path, config, refs, overwrite = FALSE)
}
\arguments{
\item{path}{character vector with folders to run the model in}

\item{config}{run configurations}

\item{refs}{named list of matching references}

\item{overwrite}{logical, should existing data be overwritten?}
}
\description{
Create data for reference matching
}
\author{
Robin Hasse
}
