% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRunFolder.R
\name{createRunFolder}
\alias{createRunFolder}
\title{Create new run folder}
\usage{
createRunFolder(
  path,
  config = NULL,
  overwrite = FALSE,
  recursive = FALSE,
  showWarnings = TRUE
)
}
\arguments{
\item{path}{character vector, containing}

\item{config}{list with run configuration}

\item{overwrite}{logical; Should exiting folders be overwritten?}

\item{recursive}{logical; Should exiting folders be overwritten?}

\item{showWarnings}{logical; Should exiting folders be overwritten?}
}
\description{
Create a folder for the model to run in and copy required gams files there.
}
\details{
Create either one or multiple
}
\author{
Robin Hasse
}
