% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.addSpecCostToInput}
\alias{.addSpecCostToInput}
\title{Write the specific costs to the input.gdx}
\usage{
.addSpecCostToInput(
  m,
  path,
  optimVarCon,
  optimVarRen,
  xinitCon,
  xinitRen,
  tcalib,
  varName = "x"
)
}
\arguments{
\item{m}{Gams transfer container with previous input data}

\item{path}{character, path to this run}

\item{optimVarCon}{data frame with optimization variables of construction}

\item{optimVarRen}{data frame with optimization variables of renovation}

\item{xinitCon}{data frame with initial specific intangible costs of construction}

\item{xinitRen}{data frame with initial specific intangible costs of renovation}

\item{tcalib}{numeric, calibration time steps}

\item{varName}{character, optimization variable to calculate specific costs from.
Needs to be a column of optimVarCon and optimVarRen, should be one of 'x', 'xA' or 'xMin'.}
}
\description{
Write the specific costs to the input.gdx
}
