% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.combineOuterObjective}
\alias{.combineOuterObjective}
\title{Read in brick results and compute the total outer objective function
by combining resuls from construction and renovation.}
\usage{
.combineOuterObjective(
  m,
  outerObjective,
  p_calibTarget,
  tcalib,
  dims,
  varName = "f",
  agg = NULL
)
}
\arguments{
\item{m}{Gams transfer container to read brick results from}

\item{outerObjective}{data frame to write the outer objective to}

\item{p_calibTarget}{data frame with historic flows}

\item{tcalib}{numeric, calibration time steps}

\item{dims}{named character, dimensions of the flows}

\item{varName}{character, column name in \code{outerObjective} to write the result to.
Should be either 'f' or 'fA'.}

\item{agg}{character, dimension(s) to aggregate the data over}
}
\description{
Read in brick results and compute the total outer objective function
by combining resuls from construction and renovation.
}
