% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.combineStepSizeParams}
\alias{.combineStepSizeParams}
\title{Combine the step size paramters delta and phi-derivative from construction and renovation flows}
\usage{
.combineStepSizeParams(paramsCon, paramsRen)
}
\arguments{
\item{paramsCon}{data frame with step size parameters from construction}

\item{paramsRen}{data frame with step size parameters from renovation}
}
\description{
Combine the step size paramters delta and phi-derivative from construction and renovation flows
}
