% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.combineStepSizeParams}
\alias{.combineStepSizeParams}
\title{Combine the step size paramters delta and phi-derivative from construction and renovation flows}
\usage{
.combineStepSizeParams(deviation)
}
\arguments{
\item{deviation}{named list of data frames with deviation and adjustment term \code{d} from all calibration variables}
}
\value{
data frame with step size adaptation parameters computed from all calibration variables
}
\description{
Combine the step size paramters delta and phi-derivative from construction and renovation flows
}
