% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.computeDeviation}
\alias{.computeDeviation}
\title{Compute the deviation between historic data and Brick results.
Then compute the adjustment term for the calibration \code{d}}
\usage{
.computeDeviation(
  m,
  target,
  dims,
  tcalib,
  flow = c("construction", "renovation"),
  renAllowed = NULL,
  vinExists = NULL,
  agg = NULL,
  calibResolution = "full"
)
}
\arguments{
\item{m}{Gams transfer container with current Brick results}

\item{target}{data frame of historical data}

\item{dims}{character, dimensions of historic data and Brick results}

\item{tcalib}{numeric, time steps to calibrate on}

\item{flow}{character, either 'construction' or 'renovation'}

\item{renAllowed}{data frame with allowed renovation transitions}

\item{vinExists}{data frame with existing vintages for each time period}

\item{agg}{character, dimensions to aggregate Brick results and target data over}

\item{calibResolution}{character, resolution of the calibration for renovation flows}
}
\description{
Compute the deviation between historic data and Brick results.
Then compute the adjustment term for the calibration \code{d}
}
