% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.computeStepSizeParams}
\alias{.computeStepSizeParams}
\title{Compute the step size adaptation paramters delta and phi-derivative}
\usage{
.computeStepSizeParams(deviation)
}
\arguments{
\item{deviation}{data frame with deviation and adjustment term 'd'}
}
\description{
Compute the step size adaptation paramters delta and phi-derivative
}
