% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.computeSumSqTarget}
\alias{.computeSumSqTarget}
\title{Compute the sum of the squares for a calibration target}
\usage{
.computeSumSqTarget(target, zeroFlow = FALSE)
}
\arguments{
\item{target}{data frame with calibration target data}

\item{zeroFlow}{logical, whether data should be filtered for the zero flow}
}
\description{
Compute the sum of the squares for a calibration target
}
