% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateScenarioRuns.R
\name{.createNewSettings}
\alias{.createNewSettings}
\title{Create new scenario bundle settings file}
\usage{
.createNewSettings(path, configFolderName = "config")
}
\arguments{
\item{path}{character, path of file to be written}

\item{configFolderName}{character, config folder to look for next to settings
file to be written. It is advised to have the config folder and the
settings file in the same directory but any sub directory relative to the
settings file is accepted.}
}
\value{
template (invisibly)
}
\description{
Writes template for settings file to given path. The file might have to be
revisited to specify place holders.
}
\author{
Robin Hasse
}
