% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.determineSpecCost}
\alias{.determineSpecCost}
\title{Assemble specific costs from initial specific costs and the optimization variable}
\usage{
.determineSpecCost(
  optimVar,
  xinit,
  dims,
  tcalib,
  flow = c("construction", "renovation", "renovationBS", "renovationHS"),
  varName = "x",
  vinExists = NULL,
  vinCalib = NULL,
  shiftIntang = TRUE
)
}
\arguments{
\item{optimVar}{data frame with the optimization variable 'x' and optionally 'xA'}

\item{xinit}{data frame with initial specific intangible costs}

\item{dims}{character, dimensions of the optimization variable}

\item{tcalib}{numeric, calibration time steps}

\item{flow}{character, type of flow, either construction or renovation}

\item{varName}{character, optimization variable to calculate specific costs from.
Should be either 'x' or 'xA'.}

\item{vinExists}{data frame of vintages that exist for each time period}

\item{vinCalib}{data frame with vintages that exist in calibration periods}

\item{shiftIntang}{logical indicating whether intangible costs should be shifted to positive range}
}
\value{
data frame with specific intangible costs for the given flow
}
\description{
Assemble specific costs from initial specific costs and the optimization variable
}
