% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateScenarioRuns.R
\name{.fileExists}
\alias{.fileExists}
\title{Check if file exists}
\usage{
.fileExists(...)
}
\value{
logical, \code{NULL} for \code{NULL}
}
\description{
Wrapper around \code{base::file.exists} with special treatment of \code{NULL}
}
