% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyHistoryGdx.R
\name{.findOriginGdxFile}
\alias{.findOriginGdxFile}
\title{find origin gdx file}
\usage{
.findOriginGdxFile(originGdx, outputFolder)
}
\arguments{
\item{originGdx}{character, file path to run or gdx file used as historical
or scenario name}

\item{outputFolder}{directory of output folder, only required if
\code{originGDX} is a scenario name}
}
\value{
file path to origin gdx file
}
\description{
Depending on what is passed via \code{originGDX}, this function returns the
file path if it exists, looks for recognised file names in the given
directory if it exists or looks for the latest run with the given name in the
\code{outputFolder}.
}
