% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.initOptimVar}
\alias{.initOptimVar}
\title{Initialize the data frame for the optimization variables}
\usage{
.initOptimVar(mInput, tcalib, dims, renAllowed)
}
\arguments{
\item{mInput}{Gamstransfer container with the input data}

\item{tcalib}{numeric, calibration time periods}

\item{dims}{character, dimensions to initialize the data with}

\item{renAllowed}{data frame with allowed renovation transitions}
}
\value{
data frame of desired dimension with x = 0 (column to contain the optimzation variable)
}
\description{
Initialize the data frame for the optimization variables
}
