% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.pick}
\alias{.pick}
\title{Pick lines from data.frame by indentifiers}
\usage{
.pick(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{<column> = <selectionElement>}
}
\value{
data.frame with filtered rows without identifier columns
}
\description{
Pick lines from data.frame by indentifiers
}
