% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.readCalibTarget}
\alias{.readCalibTarget}
\title{Read calibration targets from input folder}
\usage{
.readCalibTarget(tcalib, modifyData = FALSE)
}
\arguments{
\item{tcalib}{numeric, calibration time periods}

\item{modifyData}{logical, switch to control whether to alter data read from matching}
}
\description{
Read calibration targets from input folder
}
