% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.readCalibTarget}
\alias{.readCalibTarget}
\title{Read calibration targets from input folder}
\usage{
.readCalibTarget(tcalib)
}
\arguments{
\item{tcalib}{numeric, calibration time periods}
}
\description{
Read calibration targets from input folder
}
