% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConfig.R
\name{.readCfg}
\alias{.readCfg}
\title{Read config file}
\usage{
.readCfg(file)
}
\arguments{
\item{file}{character, path to config file}
}
\value{
named list with config parameters
}
\description{
Read yaml file from given path, check minimum requirement (title exists) and
save the file path as a attribute.
}
