% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.readOuterObjectiveOptim}
\alias{.readOuterObjectiveOptim}
\title{Read the outer objective from a gams calibration run.}
\usage{
.readOuterObjectiveOptim(m, outerObjective, varName = "f")
}
\arguments{
\item{m}{Gams transfer container to read brick results from}

\item{outerObjective}{data frame to write the outer objective to}

\item{varName}{character, column name in \code{outerObjective} to write the result to.
Should be either 'f' or 'fA'.}
}
\description{
Only applicable for optimization calibration.
}
