% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.updateStepSize}
\alias{.updateStepSize}
\title{Update the step size for the selected combinations}
\usage{
.updateStepSize(totalStep, stepSizeParams, outerObjective, stepReduction)
}
\arguments{
\item{totalStep}{data frame with combinations to perform the adjustment ot the step size on}

\item{stepSizeParams}{data frame with the parameters of the step size adjustment algorithm}

\item{outerObjective}{data frame containing the value of the outer objective function.}

\item{stepReduction}{numeric, factor applied to stepSize to reduce the step size. Should be < 1.}
}
\description{
Update the step size for the selected combinations
}
