% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{.writeCostIntang}
\alias{.writeCostIntang}
\title{Write the intangible costs to a .csv file}
\usage{
.writeCostIntang(
  file,
  optimVar,
  xinit,
  dims,
  tcalib,
  flow = c("construction", "renovation"),
  vinExists = NULL,
  vinCalib = NULL
)
}
\arguments{
\item{file}{character, path to file to which results should be written}

\item{optimVar}{data frame with optimization variable, needs to contain column 'x'}

\item{xinit}{data frame with initial intangible specific costs}

\item{dims}{character, dimensions of the optimization variable}

\item{tcalib}{numeric, calibration time steps}

\item{flow}{character, specifies flow considered. Must be cosntruction or renovation}

\item{vinExists}{data frame with existing vintage and time step combinations}

\item{vinCalib}{data frame with vintage and time step combinations that exist during calibration}
}
\description{
Write the intangible costs to a .csv file
}
