% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLatestRun.R
\name{findLatestRun}
\alias{findLatestRun}
\title{Find latest run}
\usage{
findLatestRun(run, outputFolder, fileNames = NULL)
}
\arguments{
\item{run}{character, name of a run in the output folder}

\item{outputFolder}{character, path to output folder.}

\item{fileNames}{character, file names to consider when looking for files
in run folders. If you allow multiple file names, they are tested in the
given order and the first match is returned (start with highest priority).
If left \code{NULL}, only the latest run folder is returned.}
}
\value{
path to latest run folder or the file inside this folder
}
\description{
Find the latest run of a given name in the output folder. If specified,
require files in the runs and point to them.
}
\author{
Robin Hasse
}
