% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessColnames.R
\name{guessColnames}
\alias{guessColnames}
\title{guess column names based on column values}
\usage{
guessColnames(x, m)
}
\arguments{
\item{x}{data.frame with unknown column names}

\item{m}{gams Conatiner with sets}
}
\value{
data.frame with guessed column names
}
\description{
guess column names based on column values
}
\author{
Robin Hasse
}
