% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToDf.R
\name{listToDf}
\alias{listToDf}
\title{Convert nested named list to long data.frame}
\usage{
listToDf(x, split = NULL, n = 1)
}
\arguments{
\item{x}{named list with identical depth on each branch}

\item{split}{character, split names by this pattern and assume same value for
each separated item}

\item{n}{Integer, number of nested function calls, leave default}
}
\value{
data.frame with column for each level of the named list
}
\description{
Convert nested named list to long data.frame
}
\author{
Robin Hasse
}
