% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCalibrationTarget.R
\name{loadCalibrationTarget}
\alias{loadCalibrationTarget}
\title{Load calibration targets from matching results}
\usage{
loadCalibrationTarget(config)
}
\arguments{
\item{config}{named list with run configuration}
}
\value{
directory of input folder
}
\description{
Load calibration targets from matching results
}
\author{
Ricarda Rosemann
}
