% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMadratData.R
\name{loadMadratData}
\alias{loadMadratData}
\title{Load Input data from mredgebuildings}
\usage{
loadMadratData(config)
}
\arguments{
\item{config}{named list with run configuration}
}
\value{
directory of input folder
}
\description{
Load Input data from mredgebuildings
}
\author{
Robin Hasse
}
