% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodFromConfig.R
\name{periodFromConfig}
\alias{periodFromConfig}
\title{get Period from config}
\usage{
periodFromConfig(config, which)
}
\arguments{
\item{config}{named list with run configuration}

\item{which}{type of period(s)}
}
\value{
numeric vector with periods
}
\description{
extract different periods from a given BRICK config
}
\author{
Robin Hasse
}
