% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRefDeviation.R
\name{plotRefDeviation}
\alias{plotRefDeviation}
\title{Plot heat map of reference deviation}
\usage{
plotRefDeviation(path)
}
\arguments{
\item{path}{character; directory of output folder}
}
\description{
Plot heat map of reference deviation
}
\author{
Robin Hasse
}
