% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummary.R
\name{plotSummary}
\alias{plotSummary}
\title{Plot Summary of a run}
\usage{
plotSummary(path, facet = "typ", showHistStock = FALSE, splitRen = FALSE)
}
\arguments{
\item{path}{character, path to the run}

\item{facet}{character, dimension to resolve as facets}

\item{showHistStock}{logical, plot renovation with identical replacement
transparent}
}
\description{
Plot an overview of the stock and flows
}
\author{
Robin Hasse
}
