% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConfig.R
\name{readConfig}
\alias{readConfig}
\title{Read config file}
\usage{
readConfig(config = NULL, basisOf = NULL)
}
\arguments{
\item{config}{character, config file, either a path to a yaml file or the
name of the file in `inst/config/`}

\item{basisOf}{character, name of other config that is based on this config}
}
\value{
named list with run config
}
\description{
Read config file in yaml format
}
\details{
If no argument is given, the default config is used.
}
\author{
Robin Hasse
}
