% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConfig.R
\name{readConfig}
\alias{readConfig}
\title{Read config file}
\usage{
readConfig(config = NULL, configFolder = NULL, readDirect = FALSE)
}
\arguments{
\item{config}{character, config file, either a path to a yaml file or the
name of the file in \code{configFolder}}

\item{configFolder}{character, directory to search for configs. If NULL, the
BRICK-internal config folder is used.}

\item{readDirect}{logical, specify whether \code{config} is a valid path to a
config file that should be read directly.}
}
\value{
named list with run config
}
\description{
Read config file in yaml format
}
\details{
If no argument is given, the default config is used.
}
\author{
Robin Hasse
}
