% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSlurmConfig.R
\name{setSlurmConfig}
\alias{setSlurmConfig}
\title{Set the SLURM configuration}
\usage{
setSlurmConfig(slurmQOS, tasksPerNode = 16, tasks32 = FALSE, timeLimit = NULL)
}
\arguments{
\item{slurmQOS}{string, name of the desired QOS (Quality of Service)}

\item{tasksPerNode}{numeric, number of tasks per node to be requested}

\item{tasks32}{boolean, specify whether a node with 32 tasks should be requested}

\item{timeLimit}{character, time limit of the slurm job given in the format hh:mm:ss}
}
\value{
string with SLURM configuration
}
\description{
Based on user input, construct the SLURM configuration string.
}
\details{
Check if SLURM configuration is admissible.
}
\author{
Ricarda Rosemann
}
