% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateScenarioRuns.R
\name{updateScenarioRuns}
\alias{updateScenarioRuns}
\title{Update bundle of scenario runs}
\usage{
updateScenarioRuns(bundleSettings, startFromRun = TRUE)
}
\arguments{
\item{bundleSettings}{character, path to settings file. If the file doesn't
exist, a new file is written as a template.}

\item{startFromRun}{logical, if TRUE, start from previous run saved in
settings.yaml}

\item{configFolder}{character, directory to search for configs. If NULL, the
BRICK-internal config folder is used.}

\item{outputFolder}{directory of output folder}
}
\value{
path to the started run
}
\description{
Run a bundle of scenario runs or update selected runs. This function requires
local configs and a bundle setting file. The settings file serves two
purposes: First, It holds the settings that control which scenarios should be
updated when running this function. Second, it holds the run paths to the
most recent runs that should be used by post processing, reporting or
plotting routines.
}
\details{
This function provides convenient features for scenario studies but it
assumes a specific setup that is advised to follow and might not permanent
otherwise:
\itemize{
  \item Choose one central directory from where you call this function.
  \item Place a bundle settings file in this directory. You can use this
    function to create a template.
  \item Place a config folder and an output folder inside the central
    directory and reference both as relative paths in the settings file.
}
Mind that this function changes the config files. This should only affect the
starting point. But it is advisable to commit the config files to git before
to avoid loosing information.
}
\author{
Robin Hasse
}
